Move Config to statics.rs
