+++ dynamic-tournament-web/src/lib.rs
@@ -7,7 +7,6 @@ mod services;
 mod statics;
 mod utils;
 
-use serde::{Deserialize, Serialize};
 use wasm_bindgen::prelude::*;
 use yew::start_app_in_element;
 
@@ -20,41 +19,6 @@ extern crate wee_alloc;
 #[global_allocator]
 static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
 
-/// The global config.
-///
-/// This instance always lives for the lifetime of the program.
-///
-/// # Safety
-///
-/// An `Config` instance is always expected to have a `'static` lifetime. Some methods make use of
-/// this assumption to provide safe methods.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct Config {
-    // We never need to resize this, so `Box<str>` saves us 1 * usize of space.
-    api_base: Box<str>,
-    root: Box<str>,
-}
-
-impl Config {
-    /// Returns the defined api_base.
-    #[inline]
-    pub fn api_base(&self) -> &'static str {
-        &self.static_ref().api_base
-    }
-
-    #[inline]
-    pub fn root(&self) -> &'static str {
-        &self.static_ref().root
-    }
-
-    /// Converts a `&Config` reference into a `&'static Config` reference.
-    #[inline]
-    fn static_ref(&self) -> &'static Self {
-        // SAFETY: The caller must guarantee that `self` has a `'static` lifetime.
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
 #[wasm_bindgen]
 pub fn main(config: &JsValue) {
     // SAFETY: Called from a single threaded context. No race conditions can occur.
@@ -62,7 +26,7 @@ pub fn main(config: &JsValue) {
         logger::init();
     }
 
-    let config: Config = config.into_serde().unwrap();
+    let config = config.into_serde().unwrap();
     // SAFETY: There are no references to the config.
     unsafe {
         statics::set_config(config);
