+++ dynamic-tournament-web/src/statics.rs
@@ -1,6 +1,41 @@
 use std::mem::MaybeUninit;
 
-use crate::Config;
+use serde::{Deserialize, Serialize};
+
+/// The global config.
+///
+/// This instance always lives for the lifetime of the program.
+///
+/// # Safety
+///
+/// An `Config` instance is always expected to have a `'static` lifetime. Some methods make use of
+/// this assumption to provide safe methods.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Config {
+    // We never need to resize this, so `Box<str>` saves us 1 * usize of space.
+    pub api_base: Box<str>,
+    pub root: Box<str>,
+}
+
+impl Config {
+    /// Returns the defined api_base.
+    #[inline]
+    pub fn api_base(&self) -> &'static str {
+        &self.static_ref().api_base
+    }
+
+    #[inline]
+    pub fn root(&self) -> &'static str {
+        &self.static_ref().root
+    }
+
+    /// Converts a `&Config` reference into a `&'static Config` reference.
+    #[inline]
+    fn static_ref(&self) -> &'static Self {
+        // SAFETY: The caller must guarantee that `self` has a `'static` lifetime.
+        unsafe { std::mem::transmute(self) }
+    }
+}
 
 static mut CONFIG: MaybeUninit<Config> = MaybeUninit::uninit();
